% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecostatscale.R
\name{df_col}
\alias{df_col}
\title{Simulate deterministic dynamics of N patches with dispersal}
\usage{
df_col(time, state, pars)
}
\arguments{
\item{time}{time step - ignored, but required for consistency with ode function}

\item{state}{vector of current states}

\item{pars}{parameter list, including matrix A with full interaction matrix (and growth rates along the diagonal),
and a value Ifrac, which is the dispersal rate (D in the main text), and Ksim, which is the carrying capacity}
}
\value{
rate of change for each species
}
\description{
Helper function for symdynN, used to simulate dynamics between disturbance events.
Note that K is set to 1 for all species.
}
