% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecostatscale.R
\name{xt}
\alias{xt}
\title{Simulate deterministic dynamics of a single species}
\usage{
xt(t0, t1, B0, r)
}
\arguments{
\item{t0}{initial time step}

\item{t1}{desired time step}

\item{B0}{initial abundance}

\item{r}{relative growth rate}
}
\value{
predicted value of x at time t1
}
\description{
Helper function for symdyn, used to simulate dynamics between disturbance events.
}
\examples{
# see xt2fun
}
