% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecostatscale.R
\name{xtN}
\alias{xtN}
\title{Simulate deterministic dynamics}
\usage{
xtN(t0, t1, B0, odepars, dffun, nsteps = 2)
}
\arguments{
\item{t0}{initial time step}

\item{t1}{desired time step}

\item{B0}{initial abundance}

\item{odepars}{parameter list}

\item{dffun}{function for calculating derivatives}

\item{nsteps}{number of time steps to return - defaults to 2}
}
\value{
a matrix of species abundances
}
\description{
Helper function for symdynN. Simulate an ODE given parameters, starting value, and times.
}
